Massive load of fixes
