+++ components/style.tsx
@@ -1,0 +1,18 @@
+export const divstyle = {
+    overflow: 'auto',
+    height: 'min(90vh - 200px, 1000px)',
+};
+
+export const textstyle = {
+    fontSize: '28px',
+    scrollBehavior: 'smooth' as 'smooth',
+    textAlign: 'left' as 'left',
+    lineHeight: 2,
+    letterSpacing: 1
+};
+
+export const headingStyle = {
+    fontSize: '50px',
+    scrollBehavior: 'smooth' as 'smooth',
+    textAlign: 'left' as 'left',
+};
